scripts/requirements.txt
feedparser>=6.0.0
requests>=2.31.0
google-generativeai>=0.8.0
pytrends>=4.9.2
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
